chore: implemenet CollabInteract
