+++ frontend/rust-lib/dart-ffi/Cargo.toml
@@ -29,7 +29,6 @@ appflowy-integrate = {version = "0.1.0" }
 lib-dispatch = { path = "../lib-dispatch" }
 flowy-core = { path = "../flowy-core" }
 flowy-notification = { path = "../flowy-notification" }
-flowy-net = { path = "../flowy-net" }
 flowy-derive = { path = "../../../shared-lib/flowy-derive" }
 flowy-server = { path = "../flowy-server" }
 flowy-server-config = { path = "../flowy-server-config" }
