+++ frontend/rust-lib/flowy-core/src/integrate/mod.rs
@@ -1,1 +1,4 @@
+pub(crate) mod collab_interact;
+pub(crate) mod log;
 pub(crate) mod server;
+pub(crate) mod user;
