+++ frontend/rust-lib/flowy-core/src/module.rs
@@ -18,7 +18,6 @@ pub fn make_plugins(
     .unwrap();
   let user_plugin = flowy_user::event_map::init(user_session);
   let folder_plugin = flowy_folder2::event_map::init(folder_manager);
-  let network_plugin = flowy_net::event_map::init();
   let ai_plugin = flowy_ai::event_map::init();
   let database_plugin = flowy_database2::event_map::init(database_manager);
   let document_plugin2 = flowy_document2::event_map::init(document_manager2);
@@ -26,7 +25,6 @@ pub fn make_plugins(
   vec![
     user_plugin,
     folder_plugin,
-    network_plugin,
     ai_plugin,
     database_plugin,
     document_plugin2,
