--- frontend/rust-lib/flowy-net/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "flowy-net"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lib-dispatch = { path = "../lib-dispatch" }
-protobuf = {version = "2.28.0"}
-bytes = { version = "1.4" }
-tracing = { version = "0.1"}
-
-[features]
-http_server = []
-dart = [
-    "flowy-codegen/dart",
-]
-
-ts = [
-    "flowy-codegen/ts",
-]
-
-[build-dependencies]
-flowy-codegen = { path = "../../../shared-lib/flowy-codegen"}
