--- frontend/rust-lib/flowy-net/Flowy.toml
@@ -1,3 +1,0 @@
-# Check out the FlowyConfig (located in flowy_toml.rs) for more details.
-proto_input = ["src/event_map.rs", "src/entities"]
-event_files = ["src/event_map.rs"]
