--- frontend/rust-lib/flowy-net/src/entities/mod.rs
@@ -1,2 +1,0 @@
-mod network_state;
-pub use network_state::*;
