--- frontend/rust-lib/flowy-net/src/event_map.rs
@@ -1,5 +1,0 @@
-use lib_dispatch::prelude::*;
-
-pub fn init() -> AFPlugin {
-  AFPlugin::new().name("Flowy-Network")
-}
