--- frontend/rust-lib/flowy-net/src/handlers/mod.rs
@@ -1,3 +1,0 @@
-// pub async fn update_network_ty(_data: AFPluginData<NetworkStatePB>) -> Result<(), FlowyError> {
-//   Ok(())
-// }
