--- frontend/rust-lib/flowy-net/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod entities;
-pub mod event_map;
-mod handlers;
