+++ frontend/rust-lib/flowy-user/Cargo.toml
@@ -44,6 +44,7 @@ fancy-regex = "0.11.0"
 uuid = { version = "1.3.3", features = [ "v4"] }
 chrono = { version = "0.4.22", default-features = false, features = ["clock"] }
 base64 = "^0.21"
+anyhow = "1.0.75"
 
 [dev-dependencies]
 nanoid = "0.4.0"
