+++ frontend/rust-lib/flowy-user/src/services/collab_interact.rs
@@ -1,0 +1,25 @@
+use anyhow::Error;
+pub use collab_user::core::Reminder;
+
+use lib_infra::future::FutureResult;
+
+pub trait CollabInteract: Send + Sync + 'static {
+  fn add_reminder(&self, reminder: &Reminder) -> FutureResult<(), Error>;
+  fn remove_reminder(&self, reminder_id: &str) -> FutureResult<(), Error>;
+  fn update_reminder(&self, reminder: &Reminder) -> FutureResult<(), Error>;
+}
+
+pub struct DefaultCollabInteract;
+impl CollabInteract for DefaultCollabInteract {
+  fn add_reminder(&self, _reminder: &Reminder) -> FutureResult<(), Error> {
+    FutureResult::new(async { Ok(()) })
+  }
+
+  fn remove_reminder(&self, _reminder_id: &str) -> FutureResult<(), Error> {
+    FutureResult::new(async { Ok(()) })
+  }
+
+  fn update_reminder(&self, _reminder: &Reminder) -> FutureResult<(), Error> {
+    FutureResult::new(async { Ok(()) })
+  }
+}
