+++ shared-lib/lib-infra/src/lib.rs
@@ -3,5 +3,4 @@ pub use async_trait;
 pub mod box_any;
 pub mod future;
 pub mod ref_map;
-pub mod retry;
 pub mod util;
