--- shared-lib/lib-infra/src/retry/mod.rs
@@ -1,5 +1,0 @@
-mod future;
-mod strategy;
-
-pub use future::*;
-pub use strategy::*;
