--- shared-lib/lib-infra/src/retry/strategy/jitter.rs
@@ -1,5 +1,0 @@
-use std::time::Duration;
-
-pub fn jitter(duration: Duration) -> Duration {
-  duration.mul_f64(rand::random::<f64>())
-}
