--- shared-lib/lib-infra/src/retry/strategy/mod.rs
@@ -1,7 +1,0 @@
-mod exponential_backoff;
-mod fixed_interval;
-mod jitter;
-
-pub use exponential_backoff::*;
-pub use fixed_interval::*;
-pub use jitter::*;
